Commit 20.10

+++ src/main/java/javacourse/dao/UserDao.java
@@ -10,9 +10,9 @@ public interface UserDao {
 
     void updateUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
     List<User> listUsers();
 
-    User getUserById(int id);
+    User getUserById(Long id);
 }
