+++ src/main/java/javacourse/service/UserService.java
@@ -10,9 +10,9 @@ public interface UserService {
 
     void updateUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
     List<User> listUsers();
 
-    User getUserById(int id);
+    User getUserById(Long id);
 }
